fixed the overwrite of test_selector_state
